Many big corrects

+++ rebdev.pavel/T1/outputShapes.cpp
@@ -1,9 +1,9 @@
 #include "outputShapes.hpp"
 
-void rebdev::outputShapes(Shape ** shapes, int numOfShapes, std::ostream & out)
+void rebdev::outputShapes(Shape ** shapes, size_t numOfShapes, std::ostream & out)
 {
   double sum = 0;
-  for (int i = 0; i < numOfShapes; ++i)
+  for (size_t i = 0; i < numOfShapes; ++i)
   {
     sum += shapes[i]->getArea();
   }
@@ -13,7 +13,7 @@ void rebdev::outputShapes(Shape ** shapes, int numOfShapes, std::ostream & out)
 
   out << sum;
 
-  for (int i = 0; i < numOfShapes; ++i)
+  for (size_t i = 0; i < numOfShapes; ++i)
   {
     rebdev::rectangle_t rect = shapes[i]->getFrameRect();
     rebdev::point_t lowLeft = {rect.pos.x - rect.width/2, rect.pos.y - rect.height/2};
